requirements.txt
arrow==1.3.0
asgiref==3.8.1
atomicwrites==1.4.1
attrs==24.2.0
binaryornot==0.4.4
boto3==1.27.0
botocore==1.30.1
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
cloudinary==1.41.0
cookiecutter==2.6.0
cryptography==44.0.0
decouple==0.0.7
dj-database-url==2.3.0
Django==5.1.4
django-appconf==1.0.6
django-browser-reload==1.17.0
django-cloudinary-storage==0.3.0
django-compressor==4.5.1
django-cors-headers==4.7.0
django-daraja==1.3.0
django-filter==25.1
django-heroku==0.3.1
django-jazzmin==3.0.1
django-mpesa==2.0.14
django-redis==5.4.0
django-redis-cache==3.0.1
django-resized==1.0.3
django-rest-passwordreset==1.5.0
django-storages==1.14.5
django-tailwind==3.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
ffmpeg-python==0.2.0
future==0.16.0
gunicorn==23.0.0
h11==0.14.0
idna==2.7
Jinja2==3.1.4
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.5.0
packaging==24.2
pillow==11.0.0
pluggy==0.6.0
psycopg2==2.9.10
py==1.11.0
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.18.0
PyJWT==2.10.1
pytailwindcss==0.2.0
pytest==3.6.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-magic==0.4.27
python-pesa==0.7
python-slugify==8.0.4
PyYAML==6.0.2
rcssmin==1.1.2
redis==3.5.3
requests==2.32.3
rich==13.9.4
rjsmin==1.2.2
s3transfer==0.6.2
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.2
text-unidecode==1.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.32.1
whitenoise==6.8.2
